bot/requirements.txt
python-telegram-bot[ext]==21.10
asyncpg==0.30.0
pydantic-settings==2.7.1
google-genai==1.5.0
alembic==1.14.1
psycopg2-binary==2.9.10
